fetch: spring.jpa.hibernate.ddl-auto=validateë¡ 설정변경 및l-auto=validate로  ³update and entity ㅕãmodify

+++ src/main/java/theWordI/backend/domain/meditation/entity/MeditationVerse.java
@@ -29,7 +29,7 @@ public class MeditationVerse {
     private Long verseId;
 
     @Column(name="order_no")
-    private Short orderNo;
+    private Integer orderNo;
 
     @Column(name="reg_dt", updatable = false)
     @CreatedDate
